requirements.txt
Django==5.0.4
pillow==10.3.0
django-filter==24.2
djangorestframework==3.15.1
markdown==3.6
djangorestframework-simplejwt==5.3.1
pyjwt==2.8.0
djoser==2.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-dateutil==2.9.0.post0